Use arguments avec, arho in SingleSolve
Previously, the arguments avec, arho were not used in SingleSolve, and instead,
the values in params.avec and params.arho were used.  This is presumably not the
intended behavior.

Now, this routine sets params.avec and params.arho to the arrays that are passed
in as arguments.  This is still not a good solution, as SingleSolve now has an
unexpected side effect (in particular, modifying the params that are passed in),
but it is nevertheless an improvement.